requirements.txt
django~=1.10
mysqlclient~=1.3
django-mysql~=1.1
jinja2~=2.9
django-safedelete~=0.4
django-model-utils~=2.6
six~=1.9
mariadb-dyncol~=1.2
djangorestframework~=3.5
dynamic-rest~=1.6
drf-nested-routers~=0.11